requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.5
openai==1.82.0
flask==3.0.3
twilio==9.7.2
ffmpeg==1.4